Add osg-next options to --repo
upcoming will build to the osg-upcoming-el* targets
old-upcoming will build to el*-osg-upcoming
osg will build to osg-el*
osg-M.N (or just M.N) will build to osg-M.N-el*
old-osg will build to el*-osg

This is a quick and dirty solution. The problem is that at argument parsing
time, we haven't logged into koji yet (and in fact don't know if we even will)
so we can't query which targets exist.  Refactoring is needed to fix this.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-build/branches/osgnext@17696 4e558342-562e-0410-864c-e07659590f8c