rename promotion routes: goc-el5-* -> 3.2-goc-* (SOFTWARE-1969)
a lot of the goc servers are still on osg-3.2-el6, and the current goc
promotion routes for goc-{itb,production} are 3.3 only, and the only 3.2
option is goc-el5-{itb,production} ... So in fact the one case we care
about (goc-el6-{itb,production} for osg-3.2) has no valid promotion
routes.

Since trunk is 3.3, I'm leaving goc-{itb,production} at osg33, but I'm
making 3.2-goc-{itb,production} routes for el5/el6, for use as long as
the goc uses osg-3.2.


git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-build/trunk@22015 4e558342-562e-0410-864c-e07659590f8c